+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Provide an `itoa` function. It is present in Arduino's runtime environment but not on most (all?) host systems because itoa is not a portable standard function.
+- `to_h` and `to_s` functions for `ci_config.rb`
+- `CIConfig::clone`
+- Ability to override `CIConfig` from a hash instead of just a file
+- `arduino_ci_remote.rb` now supports command line switches `--testfile-select=GLOB` and `--testfile-reject=GLOB` (which can both be repeated)
 
 ### Changed
 - Simplified the use of `Array.each` with a return statement; it's now simply `Array.find`
@@ -19,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Determining a working OSX launch command no longer breaks on non-English installations
+- `arduino_ci_remote.rb` now honors selected and rejected test files
 
 ### Security